flight_bot/requirements.txt
aiogram==3.7.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
apscheduler==3.10.4
redis[asyncio]==5.0.4
httpx==0.27.0
python-dotenv==1.0.1
